+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v5.3.0 [tbd]
+_What's new_
+* Add env var support for limiting the folders from which source files are retrieved. ([#540](https://github.com/turbot/steampipe-plugin-sdk/issues/540))
+* Add go-getter support to `TableMapFunc` - add `GetSourceFiles` method to `ConnectionMapData`. ([#542](https://github.com/turbot/steampipe-plugin-sdk/issues/542))
+
 ## v5.2.0 [2023-03-02]
 _What's new_
 * Add support for update of dynamic plugin schema based on file watching events. ([#457](https://github.com/turbot/steampipe-plugin-sdk/issues/457))